Testing attl (#22)
* Change date

We changed the date to current day and added both our initials.

* Edited the Code

We changed the moves and functions of our player, based on if our
opponent betrays us.

* Strategy #2

Incorporated second strategy on player 6.

* Strategy#3

We incorprated another strategy.

* Change Players 5,6, & 7 to Strategies

Modified players 5-7 with our strategies and implemented a back up
strategy on player 8.

* We changed the code

We change the code and moved them to the right places. We also added and
extra piece of code

* We copy and pasted our first strategy

We copied our 1st strategy and pasted it as number 4.